# Doc API/api-spec.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "API de CRUD - Uniquindio",
    "description": "Esta API permite gestionar usuarios (estudiantes y profesores) y trabajos (envíos y calificaciones).\n\nLos endpoints permiten realizar operaciones CRUD sobre usuarios y gestionar el envío y calificación de trabajos.",
    "contact": {
      "name": "Diego Alejandro Pastrana Fernandez, Santiago Fernandez, Andres Felipe Rendon y Juan Camilo Valbuena",
      "url": "https://uniquindio.edu/contact",
      "email": "diegoa.pastranaf@uqvirtual.edu.co"
    },
    "license": {
      "name": "Apache 2.0",
      "url": "https://www.apache.org/licenses/LICENSE-2.0.html"
    },
    "version": "1.0"
  },
  "externalDocs": {
    "description": "Documentación completa de la API",
    "url": "https://uniquindio.edu/api-docs"
  },
  "servers": [
    {
      "url": "http://localhost:8080/CRUD",
      "description": "Servidor Local"
    },
    {
      "url": "https://api.uniquindio.edu/CRUD",
      "description": "Servidor de Producción"
    }
  ],
  "tags": [
    {
      "name": "Trabajos",
      "description": "Endpoints para gestionar trabajos: envío y calificación."
    },
    {
      "name": "Usuarios",
      "description": "Endpoints para gestionar usuarios: creación, consulta, actualización y eliminación."
    }
  ],
  "paths": {
    "/api/usuarios/{id}": {
      "get": {
        "tags": [
          "Usuarios"
        ],
        "summary": "Obtener un usuario por ID",
        "description": "Recupera la información de un usuario dado su ID.",
        "operationId": "getUsuarioById",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "ID del usuario a consultar",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/Usuario"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "Usuarios"
        ],
        "summary": "Actualizar un usuario existente",
        "description": "Actualiza los datos de un usuario existente dado su ID.",
        "operationId": "updateUsuario",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "ID del usuario a actualizar",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Usuario"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/Usuario"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Usuarios"
        ],
        "summary": "Eliminar un usuario",
        "description": "Elimina un usuario del sistema dado su ID.",
        "operationId": "deleteUsuario",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "ID del usuario a eliminar",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/trabajos/{id}/calificar": {
      "put": {
        "tags": [
          "Trabajos"
        ],
        "summary": "Calificar un trabajo",
        "description": "Permite que un profesor califique un trabajo. Se envía un objeto JSON con el ID del profesor y la calificación.",
        "operationId": "calificarTrabajo",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "ID del trabajo a calificar",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CalificarTrabajoDTO"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/Trabajo"
                }
              }
            }
          }
        }
      }
    },
    "/api/usuarios": {
      "get": {
        "tags": [
          "Usuarios"
        ],
        "summary": "Listar todos los usuarios",
        "description": "Devuelve una lista de todos los usuarios registrados.",
        "operationId": "getAllUsuarios",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Usuario"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Usuarios"
        ],
        "summary": "Crear un nuevo usuario",
        "description": "Crea un usuario nuevo en el sistema.",
        "operationId": "createUsuario",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Usuario"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/Usuario"
                }
              }
            }
          }
        }
      }
    },
    "/api/trabajos": {
      "get": {
        "tags": [
          "Trabajos"
        ],
        "summary": "Listar todos los trabajos",
        "description": "Devuelve una lista de todos los trabajos enviados.",
        "operationId": "getAllTrabajos",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Trabajo"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Trabajos"
        ],
        "summary": "Crear un nuevo trabajo",
        "description": "Permite a un estudiante enviar un trabajo. Se requiere indicar el ID del estudiante en el objeto JSON.",
        "operationId": "createTrabajo",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Trabajo"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/Trabajo"
                }
              }
            }
          }
        }
      }
    },
    "/api/trabajos/{id}": {
      "get": {
        "tags": [
          "Trabajos"
        ],
        "summary": "Obtener un trabajo por ID",
        "description": "Recupera la información de un trabajo dado su ID.",
        "operationId": "getTrabajoById",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "ID del trabajo a consultar",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/Trabajo"
                }
              }
            }
          }
        }
      }
    },
    "/api/trabajos/profesor/{profesorId}": {
      "get": {
        "tags": [
          "Trabajos"
        ],
        "summary": "Obtener los trabajos calificados por un profesor",
        "description": "Devuelve todos los trabajos calificados por un profesor dado su ID.",
        "operationId": "getTrabajosByProfesor",
        "parameters": [
          {
            "name": "profesorId",
            "in": "path",
            "description": "ID del profesor",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Trabajo"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/trabajos/estudiante/{estudianteId}": {
      "get": {
        "tags": [
          "Trabajos"
        ],
        "summary": "Obtener los trabajos de un estudiante",
        "description": "Devuelve todos los trabajos enviados por un estudiante dado su ID.",
        "operationId": "getTrabajosByEstudiante",
        "parameters": [
          {
            "name": "estudianteId",
            "in": "path",
            "description": "ID del estudiante",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Trabajo"
                  }
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "Usuario": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "description": "Identificador único del usuario",
            "format": "int64",
            "readOnly": true,
            "example": 1
          },
          "nombre": {
            "type": "string",
            "description": "Nombre del usuario",
            "example": "Juan Pérez"
          },
          "email": {
            "type": "string",
            "description": "Correo electrónico del usuario",
            "example": "juan.perez@uniquindio.edu"
          },
          "rol": {
            "type": "string",
            "description": "Rol del usuario (PROFESOR o ESTUDIANTE)",
            "example": "ESTUDIANTE"
          },
          "clase": {
            "type": "string",
            "description": "Clase a la que pertenece el usuario",
            "example": "Programación I"
          }
        },
        "description": "Objeto JSON con los datos del nuevo usuario"
      },
      "CalificarTrabajoDTO": {
        "required": [
          "calificacion",
          "profesorId"
        ],
        "type": "object",
        "properties": {
          "profesorId": {
            "type": "integer",
            "description": "ID del profesor que califica el trabajo",
            "format": "int64",
            "example": 1
          },
          "calificacion": {
            "type": "number",
            "description": "Calificación asignada",
            "format": "double",
            "example": 95
          }
        },
        "description": "Objeto JSON que contiene el ID del profesor y la calificación"
      },
      "Trabajo": {
        "required": [
          "descripcion",
          "estudiante",
          "titulo"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "description": "Identificador único del trabajo",
            "format": "int64",
            "readOnly": true,
            "example": 1
          },
          "titulo": {
            "type": "string",
            "description": "Título del trabajo",
            "example": "Implementación de CRUD en Spring Boot"
          },
          "descripcion": {
            "type": "string",
            "description": "Descripción o contenido del trabajo (por ejemplo, código o explicación detallada)",
            "example": "Este trabajo muestra la implementación de un CRUD con Spring Boot..."
          },
          "calificacion": {
            "type": "number",
            "description": "Calificación asignada por el profesor. Puede ser nula si aún no se ha calificado.",
            "format": "double",
            "nullable": true,
            "example": 85
          },
          "fechaEnvio": {
            "type": "string",
            "description": "Fecha y hora en que se envió el trabajo",
            "format": "date-time",
            "readOnly": true
          },
          "estudiante": {
            "$ref": "#/components/schemas/Usuario"
          },
          "profesor": {
            "$ref": "#/components/schemas/Usuario"
          }
        },
        "description": "Objeto JSON con los datos del trabajo a crear (debe incluir el ID del estudiante)"
      }
    }
  }
}